+++ packages/core/src/types/common/i18n/index.ts
@@ -3,3 +3,4 @@ export * from './trustee';
 export * from './actuary';
 export * from './corporateGroup';
 export * from './independentTrustee';
+export * from './inHouse';
