+++ packages/core/src/types/common/providers/index.ts
@@ -3,3 +3,4 @@ export * from './trustee';
 export * from './actuary';
 export * from './corporateGroup';
 export * from './independentTrustee';
+export * from './inHouse';
