+++ packages/layout/src/components/__tests__/inHouse.spec.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { InHouseCard } from '../cards/inHouse/inHouse';
 import { axe } from 'jest-axe';
 import userEvent from '@testing-library/user-event';
-import { InHouseAdminNoApi } from '../cards/inHouse/context';
+import { InHouseAdminNoApi } from '@tpr/core';
 
 // TODO: write more tests
