+++ packages/layout/src/components/cards/inHouse/views/remove/confirm/confirm.tsx
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
 import { useInHouseAdminContext } from '../../../context';
-import { InHouseAdminI18nProps } from '../../../i18n';
 import {
 	Breadcrumbs,
 	BreadcrumbLink,
 } from '../../../../components/breadcrumbs';
 import Confirm from '../../../../common/views/remove/confirm/confirm';
-import { cardType, cardTypeName } from '../../../../common/interfaces';
+import { cardType, cardTypeName, InHouseAdminI18nProps } from '@tpr/core';
 
 const getBreadcrumbLinks = (
 	labels: Partial<InHouseAdminI18nProps['remove']['confirm']['breadcrumbs']>,
